+++ src/index.ts
@@ -91,6 +91,8 @@ export { tsAdapter, MODULE_ANCHOR_NAME, classifyTsFile } from "./lib/ts-adapter.
 export type { TsAnchorKind, TsFileMode, TsClassification } from "./lib/ts-adapter.js";
 export { pyAdapter, classifyPyFile, warmPythonAdapter } from "./lib/py-adapter.js";
 export type { PyAnchorKind, PyFileMode, PyClassification } from "./lib/py-adapter.js";
+export { goAdapter, classifyGoFile, warmGoAdapter } from "./lib/go-adapter.js";
+export type { GoAnchorKind, GoFileMode, GoClassification } from "./lib/go-adapter.js";
 export { sfcAdapter, classifySfcFile } from "./lib/sfc-adapter.js";
 export type { SfcFileMode, SfcClassification } from "./lib/sfc-adapter.js";
 export { resolveOwner, splitAnchorId } from "./lib/ownership.js";